\name{MetaNetAssoc-package}
\alias{MetaNetAssoc}
\alias{MetaNetAssoc-package}
\docType{package}
\title{Metabolite to Biological Term Association through MPI and PPI Networks}
\description{
MetaNetAssoc connects candidate metabolites to enriched biological terms through
metabolite-protein interaction (MPI) and protein-protein interaction (PPI)
networks. It preserves supporting paths and reports raw, weighted, normalized,
and coverage scores.
}
\details{
Scores describe network-supported association potential and should not be
interpreted as evidence of causal regulation.
}
\author{MetaNetAssoc Team}

